requirements.txt
streamlit
ultralytics
Pillow
streamlit-drawable-canvas